Added anility to keep unzipped directories

+++ src/lib/InstallerControllerExtensions.php
@@ -39,6 +39,8 @@ class InstallerControllerExtensions extends BaseController
 		$params       = new Registry(JPluginHelper::getPlugin('system', 'extensionexport')->params);
 		$exportPath   = JPATH_ROOT . '/' . $params->get('directory', 'images/dist');
 		$fileMode     = (int) octdec('0' . $params->get('filemode', '644'));
+		$dirMode      = (int) octdec('0' . $params->get('dirmode', '755'));
+		$delDirs      = (bool) (int) $params->get('del_dirs', '1');
 
 		$doCreatePackage = count($extensionIds) > 1;
 		$packageData     = [];
@@ -55,10 +57,15 @@ class InstallerControllerExtensions extends BaseController
 
 			try
 			{
-				$exporter = new Exporter($exportPath);
+				$exporter = new Exporter($exportPath, $dirMode, $fileMode);
 				$package  = $exporter->export($element, $type, $clientId, $pluginGroup);
 				chmod($exportPath . '/' . $package . '.zip', $fileMode);
 				$packageData[$extensionId]->filename = $package . '.zip';
+
+				if ($delDirs)
+				{
+					$exporter->removeArtifacts();
+				}
 
 				$this->issueSuccessMessage($package, $type, $exportPath);
 			}
@@ -69,7 +76,7 @@ class InstallerControllerExtensions extends BaseController
 				$this->issueFailureMessage($element, $type, $exception);
 			}
 
-			if (!empty($package))
+			if ($delDirs && !empty($package))
 			{
 				Folder::delete($exportPath . '/' . preg_replace('~-[\d\.]*$~', '', $package));
 			}
